+++ crates/jni/tests/builtin_bindings.rs
@@ -7,7 +7,7 @@ fn test_builtin_bindings() {
     let jvm = util::jvm();
 
     jvm.attach_current_thread(|env| -> jni::errors::Result<()> {
-        jni::objects::_test_jni_init(env, &Default::default());
+        jni::__test_bindings_init(env, &Default::default());
         Ok(())
     })
     .unwrap();